Fix for the decimal point key on the numpad not working in IE
It seems that the decimal point key on the numpad has `e.key === 'delete'` rather than `'.'` for some reason, so I've made it rely purely on the charcode which seems more reliable. I'm not sure why it was using e.key sometimes rather than the charcode, so this will need properly reviewing.